requirements.txt
tensorflow==2.8.0
numpy==1.22.3
Flask==2.0.3
keras==2.8.0
Pillow==9.0.1